+++ src/hisp/__init__.py
@@ -2,4 +2,4 @@ from .helpers import PulsedSource, Scenario
 
 from .h_transport_class import CustomProblem
 
-from .festim_mb_model import make_mb_model
+from .festim_models.mb_model import make_mb_model
