+++ src/hisp/festim_models/__init__.py
@@ -1,0 +1,1 @@
+from .mb_model import make_mb_model
